+++ packages/terra-alert/src/terra-dev-site/test/alert/AlertPadding.test.jsx
@@ -1,0 +1,4 @@
+import React from 'react';
+import Alert from '../../../Alert';
+
+export default () => (<Alert id="errorAlert" type={Alert.Opts.Types.ERROR}>a b c d e f g h i j k l m n o p q r s t u v w x y z a b c d e f g h i j k l m n o p q r s t u v w x y z a b c d e f g h i j k l m n o p q r s t u v w x y z a b c d e f g h i j k l m n o p q r s t u v w x y z</Alert>);